make a bidder

+++ auctions/models.py
@@ -6,7 +6,7 @@ class User(AbstractUser):
 
 class Listing(models.Model):
     name = models.CharField(max_length=64)
-    starting_bid = models.DecimalField(max_digits=16, decimal_places=2)
+    starting_bid = models.IntegerField()
     description = models.CharField(max_length=254)
     url = models.CharField(max_length=254)
     date = models.DateTimeField(auto_now_add=True, blank=True)
@@ -24,7 +24,7 @@ class Watchlist(models.Model):
 class Bid(models.Model):
     user_bid = models.ForeignKey(User, on_delete=models.CASCADE, related_name='user_bid')
     item_bid = models.ForeignKey(Listing, on_delete=models.CASCADE, related_name='item_bid')
-    bid = models.DecimalField(max_digits=16, decimal_places=2)
+    bid = models.IntegerField()
 
     def __str__(self):
         return f"{self.user_bid} {self.item_bid} {self.bid}"
